refactor: organize components into dedicated folders
- Move FileViewer components to src/components/FileViewer/
  - FileViewer.tsx -> FileViewer/index.tsx
  - VirtualizedTextViewer.tsx -> FileViewer/VirtualizedTextViewer.tsx
  - MediaViewer.tsx -> FileViewer/MediaViewer.tsx
  - ArchiveViewer.tsx -> FileViewer/ArchiveViewer.tsx
  - VirtualizedArchiveList.tsx -> FileViewer/VirtualizedArchiveList.tsx

- Move FileBrowser components to src/components/FileBrowser/
  - FileBrowser.tsx -> FileBrowser/index.tsx
  - VirtualizedFileList.tsx -> FileBrowser/VirtualizedFileList.tsx
  - PerformanceIndicator.tsx -> FileBrowser/PerformanceIndicator.tsx
  - SettingsPanel.tsx -> FileBrowser/SettingsPanel.tsx

- Update import paths to use relative references (../../)
- Remove obsolete VirtualizedText.tsx component
- Maintain shared components in original locations (LanguageSwitcher, common/)

This improves code organization by grouping related components together
while preserving all functionality and maintaining clean import paths.
