+++ src/components/FileBrowser/PerformanceIndicator.tsx
